Bug fixes and Unit test project
Fixed the issue of an answer staying correct regardless of what language of the day a user is on. Wrote a function to test if buttons should be enabled or not based upon the position of the dictionary the user is viewing. Added a UnitTest project and a basic test to check file parsing.