Page swapping function, Connection managing
Solution now detects if the user is not connected to the internet and if so pushes a new page onto the stack which informs the user they are not connected. I implemented this as I am planning on reading the tips of the day from the internet in the future.