Added a pop-up in app
Converted ConnectivityPage into a Popup using the xamarin community toolkit. Instead of a new page being pushed onto the stack it is now a simple popup that the user can leave by tapping outside the box or clicking the exit button. It informs the user that no new tips will be added while they are offline.